+++ src/Tests/IdealIssuersTest.php
@@ -8,11 +8,16 @@ use PHPUnit\Framework\TestCase;
 
 class IdealIssuersTest extends TestCase
 {
-    private IdealIssuers $issuers;
-
-    public function setUp(): void
+    public function test_item_type()
     {
-        $this->issuers = new IdealIssuers(
+        self::assertSame(
+            expected: Issuer::class,
+            actual: IdealIssuers::ITEM_TYPE
+        );
+    }
+    public function test_to_array()
+    {
+        $issuers = new IdealIssuers(
             new Issuer(
                 id: 'ak12',
                 listType: 'admin',
@@ -25,10 +30,6 @@ class IdealIssuersTest extends TestCase
             ),
 
         );
-    }
-
-    public function test_to_array()
-    {
         self::assertSame(
             expected: [
                 [
@@ -42,20 +43,35 @@ class IdealIssuersTest extends TestCase
                     "name" => "deposit"
                 ]
             ],
-            actual: $this->issuers->toArray()
+
+            actual: $issuers->toArray()
         );
     }
 
     public function test_get_property_name()
     {
+        $issuers = new IdealIssuers(
+            new Issuer(
+                id: 'ak12',
+                listType: 'admin',
+                name: 'bill'
+            ),
+            new Issuer(
+                id: 'fp11',
+                listType: 'custom',
+                name: 'deposit'
+            ),
+
+        );
         self::assertSame(
             expected: 'issuers',
-            actual: $this->issuers->getPropertyName()
+            actual: $issuers->getPropertyName()
         );
     }
 
     public function test_add_issuer()
     {
+        $issuers = new IdealIssuers();
         self::assertEqualsCanonicalizing(
             [
                 'id' => '1',
@@ -63,16 +79,29 @@ class IdealIssuersTest extends TestCase
                 'list_type' => 'test'
 
             ],
-            $this->issuers->addIssuer(item: new Issuer(
+            $issuers->addIssuer(item: new Issuer(
                 id: '1',
                 listType: 'test',
                 name: 'test_issuer'
-            ))->get(3)->toArray()
+            ))->get()->toArray()
         );
     }
 
     public function test_remove_issuer()
     {
+        $issuers = new IdealIssuers(
+            new Issuer(
+                id: 'ak12',
+                listType: 'admin',
+                name: 'bill'
+            ),
+            new Issuer(
+                id: 'fp11',
+                listType: 'custom',
+                name: 'deposit'
+            ),
+
+        );
         self::assertEqualsCanonicalizing(
             [
                 [
@@ -81,7 +110,7 @@ class IdealIssuersTest extends TestCase
                     "name" => "bill"
                 ]
             ],
-            $this->issuers->removeIssuer(2)->toArray()
+            $issuers->removeIssuer(1)->toArray()
         );
     }
 }